requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
Deprecated==1.2.14
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
idna==3.4
importlib-resources==6.0.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
MarkupSafe==2.1.3
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
pkgutil_resolve_name==1.3.10
pycparser==2.21
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wrapt==1.15.0
zipp==3.16.2
